+++ src/client/app/components/root.tsx
@@ -17,16 +17,18 @@ const AppContainer = styled.div`
   overflow-y: auto;
 `;
 
-// const StyledNav = styled.nav`
-//   background-color: white;
-//   height: max-content;
-// `;
+const StyledNav = styled.nav`
+  background-color: white;
+  height: max-content;
+`;
+
+const basePath = '/';
 
 export const Root = () => {
   return (
     <AppContainer>
-      {/* <Router> */}
-      {/* <StyledNav>
+      <Router>
+        <StyledNav>
           <ul>
             <li>
               <Link to='/login'>Login</Link>
@@ -38,20 +40,14 @@ export const Root = () => {
               <Link to='/leaderboard'>Leaderboard</Link>
             </li>
           </ul>
-        </StyledNav> */}
+        </StyledNav>
 
-      {/* <Switch> */}
-      {/* <Route path="/">
-          <MainLeaderboard />
-        </Route> */}
-      {/* <Route path='/leaderboard'> */}
-      <MainLeaderboard />
-      {/* </Route>
-          <Route path='/team'>
-            <TeamPage />
-          </Route> */}
-      {/* </Switch> */}
-      {/* </Router> */}
+        <Switch>
+          <Route exact path='/' component={MainLeaderboard} />
+          <Route exact path='/leaderboard' component={MainLeaderboard} />
+          <Route exact path='/team' component={TeamPage} />
+        </Switch>
+      </Router>
     </AppContainer>
   );
 };
